requirements.txt
asyncpg==0.20.0
fastapi==0.46.0
uvicorn==0.11.1
pytest==5.3.2
requests==2.22.0
yapf
PyYAML
pymongo
http
boto3